+++ docs/CHANGELOG.md
@@ -41,3 +41,5 @@
 - Use RelateNG instead of Relate algorithm
 - Add missing implementation on PostGIS datasource wrapper (unwrap and iswrapper)
 - Use connection isWrapperFor instead of static class package search for checking if connection is H2 or POSTGIS
+- Fix timestamp and data IO with FlatGeoBuffer
+- Fix writing varchar with FlatGeoBuffer